requirements-full.txt
# Vanta Requirements - Full Development Environment
# This file includes ALL dependencies for local development
# Includes ML libraries, testing tools, and development utilities
# Use this for local development: pip install -r requirements-full.txt

# Core dependencies
setuptools>=65.5.1
numpy<2.0.0  # Fixed compatibility issue with PyTorch
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
transformers>=4.5.0
torch>=1.9.0
matplotlib>=3.7.2

# Web framework
flask>=2.0.0
flask-cors>=3.0.10
Jinja2>=3.1.0
flask-sqlalchemy>=3.1.1
flask-login>=0.5.0
flask-migrate>=3.0.0
psycopg2-binary>=2.9.9  # PostgreSQL driver for production deployments (Railway/Supabase)

# GUI
PyQt5>=5.15.0
PyQt6-Charts>=6.2.0
qtawesome>=1.1.0

# Data handling
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dateutil>=2.8.0
pytz>=2021.1

# Testing
pytest>=6.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-html>=3.2.0

# Production server
gunicorn>=20.1.0

# Environment variables
python-dotenv>=0.19.0

# Security
Flask-WTF>=1.0.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.60.0
colorlog>=6.6.0

# API integrations
tweepy>=4.0.0
facebook-sdk==3.1.0

# Database
SQLAlchemy>=2.0.16  # Compatible with flask-sqlalchemy 3.1.1
alembic>=1.12.0

# Code quality
black>=21.5b0
mypy>=1.5.0
flake8>=3.9.0
isort>=5.8.0

# NLP and ML dependencies - use pre-compiled wheels
spacy==3.7.2  # Version compatible with Python 3.12
# Use a compatible spacy model
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
textstat>=0.7.3
PyJWT==2.6.0
bcrypt==4.0.1
Pillow>=8.0.0

# Web interface dependencies
bootstrap-flask>=2.3.0

# Task queue
celery>=5.1.0
redis>=4.0.0
